Update11/12

+++ HS/HS.ERP/HS.ERP.DataBase/HS.ERP.DataBase_Version_1.1/dbo/Tables/Contact.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[Contact]
 (
-	[ContactId] INT NOT NULL , 
-    [AccountId] INT NOT NULL REFERENCES [AccountInfo]([AccountId]), 
+	[ContactId] BIGINT NOT NULL , 
+    [AccountId] BIGINT NOT NULL REFERENCES [AccountInfo]([AccountId]), 
     [ContactName] NVARCHAR(20) NOT NULL, 
     [Department] NVARCHAR(30) NULL, 
     [Position] NCHAR(10) NULL,
