+++ HS/HS.ERP/HS.ERP.DataBase/HS.ERP.DataBase_Version_1.1/dbo/Tables/Order.sql
@@ -1,11 +1,11 @@
 ﻿CREATE TABLE [dbo].[Order]
 (
 	[OrderId] BIGINT NOT NULL , 
-    [AccountId] INT NOT NULL, 
+    [AccountId] BIGINT NOT NULL, 
     [OrderPrice] MONEY NOT NULL, 
     [Status] BIT NOT NULL, 
     [CreatedDate] TIMESTAMP NOT NULL
-
-	CONSTRAINT [PK_OrderId] PRIMARY KEY ([OrderId]), 
+	
+	CONSTRAINT [PK_OrderId] PRIMARY KEY ([OrderId]),    
     CONSTRAINT [FK_AccountId] FOREIGN KEY ([AccountId]) REFERENCES [AccountInfo]([AccountId])
 )
